[gh-#1250] refactor and cli reorg
