+++ packages/cli/lib/constants.ts
@@ -1,0 +1,4 @@
+export const TYPES_FILE_NAME = 'models.ts';
+export const NangoSyncTypesFileLocation = 'dist/nango-sync.d.ts';
+export const NANGO_INTEGRATIONS_NAME = 'nango-integrations';
+export const exampleSyncName = 'github-issue-example';
