+++ packages/cli/lib/services/yaml.service.ts
@@ -1,9 +1,9 @@
-import type { SimplifiedNangoIntegration } from '@nangohq/shared';
+import type { StandardNangoConfig } from '@nangohq/shared';
 import { loadSimplifiedConfig, nangoConfigFile } from '@nangohq/shared';
 import { printDebug } from '../utils.js';
 
 class YamlService {
-    public async getConfig(optionalLoadLocation = '', debug = false): Promise<SimplifiedNangoIntegration[]> {
+    public async getConfig(optionalLoadLocation = '', debug = false): Promise<StandardNangoConfig[]> {
         const loadLocation = optionalLoadLocation || './';
         const config = await loadSimplifiedConfig(loadLocation);
 
@@ -18,7 +18,7 @@ class YamlService {
         return config;
     }
 
-    public getModelNames(config: SimplifiedNangoIntegration[]): string[] {
+    public getModelNames(config: StandardNangoConfig[]): string[] {
         const modelNames = config.reduce((acc: string[], config) => {
             const syncs = config.syncs || [];
             const actions = config.actions || [];
